+++ plugins/implementation/chart2/src/main/java/ChartPlugin2Page.java
@@ -32,11 +32,11 @@ public class ChartPlugin2Page extends JPanel {
         // CHART DATA
         DefaultPieDataset dataset = new DefaultPieDataset();
 
-        Integer[] monthValue = new Integer[12];
+        Double[] monthValue = new Integer[12];
         for (int i = 0; i < 12; i++) {
             monthValue[i] = 0;
         }
-    
+
         ArrayList<FixedBill> bills = dataService.getFixedBills().getElements();
         for (FixedBill bill : bills) {
             String[] date = bill.getDate().split("-");
